refactor: extract AI response and model helpers into src/sync/utils.ts
Move duplicated helper functions out of commit.ts and service.ts into a shared utils file and update callers to import , , and . Add error logging to the AI analysis path.
